Made edits after review.
