+++ src/pages/AuthPages/AuthenticationPage/AuthenticationPage.module.scss
@@ -1,61 +1,69 @@
 .header {
-  position: relative;
+  margin-bottom: 32px;
 
-  margin-top: 200px;
-
-  .icon_login {
-    width: 72px;
-
-    position: absolute;
-    top: -100px;
-    left: 0px;
-  }
   .title_header {
     margin-bottom: 3px;
 
     color: rgb(82, 82, 82);
+    font-family: Nunito Sans;
     font-size: 36px;
     font-weight: 700;
+  }
+  .paragraph_header {
+    color: rgb(82, 82, 82);
+    font-family: Nunito Sans;
+    font-size: 16px;
+    font-weight: 400;
   }
 }
 
 .input_label {
-  margin-top: 14px;
-  margin-bottom: 4px;
   color: rgb(130, 130, 130);
+  font-family: Nunito Sans;
   font-size: 14px;
   font-weight: 600;
   text-align: left;
 }
 
 .input_field {
-  width: 420px;
   height: 45px;
+  width: 420px;
+
+  font-family: Nunito Sans;
 }
 
 .button_login {
+  margin-top: 56px;
+
   width: 420px;
   height: 50px;
 
   color: rgb(255, 255, 255);
+  font-family: Nunito Sans;
   font-size: 18px;
   font-weight: 800;
   background: rgb(127, 38, 91);
 }
 
 .footer {
-  margin-top: 100px;
+  margin-top: 185px;
   margin-bottom: 75px;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
 
   .footer_paragraph {
     color: rgb(130, 130, 130);
+    font-family: Nunito Sans;
     font-size: 18px;
     font-weight: 400;
-    text-align: center;
   }
   .footer_link {
-
     color: rgb(127, 38, 91);
+    font-family: Nunito Sans;
     font-size: 18px;
     font-weight: 600;
   }
