--- src/pages/AuthenticationPage/AuthenticationPage.module.scss
@@ -1,119 +1,0 @@
-.header {
-  position: relative;
-
-  margin-top: 278px;
-
-  .icon_login {
-    width: 72px;
-
-    position: absolute;
-    top: -100px;
-    left: 0px;
-  }
-  .title_header {
-    margin-bottom: 3px;
-
-    color: rgb(82, 82, 82);
-    font-size: 36px;
-    font-weight: 700;
-  }
-  .paragraph_header {
-    margin-top: 0px;
-
-    color: rgb(82, 82, 82);
-    font-size: 16px;
-    font-weight: 400;
-  }
-}
-
-.button_google {
-  margin-top: 32px;
-
-  height: 45px;
-  width: 420px;
-
-  color: rgb(130, 130, 130);
-  font-size: 14px;
-  font-weight: 700;
-  .icon_google {
-    width: 25px;
-  }
-}
-
-.link_block {
-  margin-top: 36px;
-  margin-bottom: 12px;
-  text-align: center;
-  .link_google {
-    text-decoration: none;
-    font-size: 12px;
-    font-weight: 600;
-    color: rgb(161, 161, 161);
-    .dash_dach {
-      color: rgb(221, 221, 221);
-    }
-  }
-}
-
-.input_label {
-  margin-top: 24px;
-  margin-bottom: 4px;
-  color: rgb(130, 130, 130);
-  font-size: 14px;
-  font-weight: 600;
-  text-align: left;
-}
-
-.input_field {
-  height: 45px;
-  width: 420px;
-}
-
-.checkbox_block {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  .checkbox_text {
-    color: rgb(161, 161, 161);
-    font-size: 12px;
-    font-weight: 400;
-  }
-  .forgot_link {
-    text-decoration: none;
-    color: rgb(127, 38, 91);
-    font-size: 12px;
-    font-weight: 600;
-  }
-}
-
-.button_login {
-  width: 420px;
-  height: 50px;
-
-  color: rgb(255, 255, 255);
-  font-size: 18px;
-  font-weight: 800;
-  background: rgb(127, 38, 91);
-}
-
-.footer {
-  margin-top: 185px;
-  margin-bottom: 75px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  .footer_paragraph {
-    color: rgb(130, 130, 130);
-    font-size: 18px;
-    font-weight: 400;
-  }
-  .footer_link {
-    color: rgb(127, 38, 91);
-    font-size: 18px;
-    font-weight: 600;
-  }
-}
