+++ src/store/hook.ts
@@ -1,4 +1,8 @@
-import { useDispatch, useSelector, type TypedUseSelectorHook } from "react-redux";
+import {
+  useDispatch,
+  useSelector,
+  type TypedUseSelectorHook,
+} from "react-redux";
 import type { AppDispatch, RootState } from "./index";
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
